Put Verbose in it's own file
